review fixes: extract master restart related stuff to another files
